Bugfix: Pan works correct now (Thanks to Scott Bahling)
git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@237 3b3fb362-3133-0410-aa15-cf69e0a59cb7

+++ gui/src/Mixer/MixerLine.h
@@ -103,8 +103,8 @@ class MixerLine: public PixmapWidget
 		void setName(QString name) {     m_pNameWidget->setText( name );        }
 		QString getName() {      return m_pNameWidget->text();        }
 
-		int getPan();
-		void setPan(int value);
+		float getPan();
+		void setPan(float value);
 
 		int getActivity() {	return m_nActivity;	}
 		void setActivity( uint value ) {	m_nActivity = value;	}
